Windows, JNI, refactor: extract process mgmt logic
Refactor the Windows JNI library:

- create process.cc and process.h
- create a WaitableProcess class
- move the process creation, waiting, and
  termination logic to WaitableProcess

Next step: use WaitableProcess in the
Windows-native test wrapper, to let it kill the
subprocess tree and fix https://github.com/bazelbuild/bazel/issues/8005
