+++ src/main/native/windows/processes-jni.cc
@@ -28,13 +28,8 @@
 #include "src/main/native/jni.h"
 #include "src/main/native/windows/file.h"
 #include "src/main/native/windows/jni-util.h"
+#include "src/main/native/windows/process.h"
 #include "src/main/native/windows/util.h"
-
-// These are the possible return values from the NativeProcess::WaitFor()
-// method.
-static const int kWaitSuccess = 0;
-static const int kWaitTimeout = 1;
-static const int kWaitError = 2;
 
 template <typename T>
 static std::wstring ToString(const T& e) {
@@ -157,10 +152,6 @@ class NativeProcess {
       : stdin_(INVALID_HANDLE_VALUE),
         stdout_(),
         stderr_(),
-        process_(INVALID_HANDLE_VALUE),
-        job_(INVALID_HANDLE_VALUE),
-        ioport_(INVALID_HANDLE_VALUE),
-        exit_code_(STILL_ACTIVE),
         error_(L"") {}
 
   ~NativeProcess() {
@@ -170,48 +161,17 @@ class NativeProcess {
 
     stdout_.Close();
     stderr_.Close();
-
-    if (process_ != INVALID_HANDLE_VALUE) {
-      CloseHandle(process_);
-    }
-
-    if (job_ != INVALID_HANDLE_VALUE) {
-      CloseHandle(job_);
-    }
-
-    if (ioport_ != INVALID_HANDLE_VALUE) {
-      CloseHandle(ioport_);
-    }
   }
 
   jboolean Create(JNIEnv* env, jstring java_argv0, jstring java_argv_rest,
                   jbyteArray java_env, jstring java_cwd,
                   jstring java_stdout_redirect, jstring java_stderr_redirect,
                   jboolean redirectErrorStream) {
-    std::wstring argv0;
     std::wstring wpath(bazel::windows::GetJavaWpath(env, java_argv0));
-    std::wstring error_msg(
-        bazel::windows::AsExecutablePathForCreateProcess(wpath, &argv0));
-    if (!error_msg.empty()) {
-      error_ = bazel::windows::MakeErrorMessage(
-          WSTR(__FILE__), __LINE__, L"nativeCreateProcess", wpath, error_msg);
-      return false;
-    }
-
-    std::wstring commandline =
-        argv0 + L" " + bazel::windows::GetJavaWstring(env, java_argv_rest);
     std::wstring stdout_redirect = bazel::windows::AddUncPrefixMaybe(
         bazel::windows::GetJavaWpath(env, java_stdout_redirect));
     std::wstring stderr_redirect = bazel::windows::AddUncPrefixMaybe(
         bazel::windows::GetJavaWpath(env, java_stderr_redirect));
-    std::wstring cwd;
-    std::wstring wcwd(bazel::windows::GetJavaWpath(env, java_cwd));
-    error_msg = bazel::windows::AsShortPath(wcwd, &cwd);
-    if (!error_msg.empty()) {
-      error_ = bazel::windows::MakeErrorMessage(
-          WSTR(__FILE__), __LINE__, L"nativeCreateProcess", wpath, error_msg);
-      return false;
-    }
 
     const bool stdout_is_stream = stdout_redirect.empty();
     const bool stderr_is_stream =
@@ -222,11 +182,6 @@ class NativeProcess {
          stderr_redirect.size() == stdout_redirect.size() &&
          _wcsnicmp(stderr_redirect.c_str(), stdout_redirect.c_str(),
                    stderr_redirect.size()) == 0);
-
-    std::unique_ptr<WCHAR[]> mutable_commandline(
-        new WCHAR[commandline.size() + 1]);
-    wcsncpy(mutable_commandline.get(), commandline.c_str(),
-            commandline.size() + 1);
 
     SECURITY_ATTRIBUTES sa = {0};
     sa.nLength = sizeof(SECURITY_ATTRIBUTES);
@@ -240,8 +195,6 @@ class NativeProcess {
     bazel::windows::AutoHandle stdout_process;
     bazel::windows::AutoHandle stderr_process;
     bazel::windows::AutoHandle thread;
-    PROCESS_INFORMATION process_info = {0};
-    JOBOBJECT_EXTENDED_LIMIT_INFORMATION job_info = {0};
 
     JavaByteArray env_map(env, java_env);
     if (env_map.ptr() != nullptr) {
@@ -365,225 +318,35 @@ class NativeProcess {
       stderr_process = pipe_write_h;
     }
 
-    // MDSN says that the default for job objects is that breakaway is not
-    // allowed. Thus, we don't need to do any more setup here.
-    HANDLE job = CreateJobObject(NULL, NULL);
-    if (job == NULL) {
-      DWORD err_code = GetLastError();
-      error_ = bazel::windows::MakeErrorMessage(
-          WSTR(__FILE__), __LINE__, L"nativeCreateProcess", wpath, err_code);
-      return false;
-    }
-
-    job_ = job;
-
-    job_info.BasicLimitInformation.LimitFlags =
-        JOB_OBJECT_LIMIT_KILL_ON_JOB_CLOSE;
-    if (!SetInformationJobObject(job, JobObjectExtendedLimitInformation,
-                                 &job_info, sizeof(job_info))) {
-      DWORD err_code = GetLastError();
-      error_ = bazel::windows::MakeErrorMessage(
-          WSTR(__FILE__), __LINE__, L"nativeCreateProcess", wpath, err_code);
-      return false;
-    }
-
-    HANDLE ioport = CreateIoCompletionPort(INVALID_HANDLE_VALUE, nullptr, 0, 1);
-    if (ioport == nullptr) {
-      DWORD err_code = GetLastError();
-      error_ = bazel::windows::MakeErrorMessage(
-          WSTR(__FILE__), __LINE__, L"nativeCreateProcess", wpath, err_code);
-      return false;
-    }
-    ioport_ = ioport;
-    JOBOBJECT_ASSOCIATE_COMPLETION_PORT port;
-    port.CompletionKey = job;
-    port.CompletionPort = ioport;
-    if (!SetInformationJobObject(job,
-                                 JobObjectAssociateCompletionPortInformation,
-                                 &port, sizeof(port))) {
-      DWORD err_code = GetLastError();
-      error_ = bazel::windows::MakeErrorMessage(
-          WSTR(__FILE__), __LINE__, L"nativeCreateProcess", wpath, err_code);
-      return false;
-    }
-
-    std::unique_ptr<bazel::windows::AutoAttributeList> attr_list;
-    if (!bazel::windows::AutoAttributeList::Create(
-            stdin_process, stdout_process, stderr_process, &attr_list,
-            &error_msg)) {
-      error_ = bazel::windows::MakeErrorMessage(
-          WSTR(__FILE__), __LINE__, L"nativeCreateProcess", L"", error_msg);
-      return false;
-    }
-
-    // kMaxCmdline value: see lpCommandLine parameter of CreateProcessW.
-    static constexpr size_t kMaxCmdline = 32767;
-
-    std::wstring cmd_sample = mutable_commandline.get();
-    if (cmd_sample.size() > 200) {
-      cmd_sample = cmd_sample.substr(0, 195) + L"(...)";
-    }
-    if (wcsnlen_s(mutable_commandline.get(), kMaxCmdline) == kMaxCmdline) {
-      std::wstringstream error_msg;
-      error_msg << L"command is longer than CreateProcessW's limit ("
-                << kMaxCmdline << L" characters)";
-      error_ = bazel::windows::MakeErrorMessage(
-          WSTR(__FILE__), __LINE__, L"CreateProcessWithExplicitHandles",
-          cmd_sample, error_msg.str().c_str());
-      return false;
-    }
-
-    STARTUPINFOEXW info;
-    attr_list->InitStartupInfoExW(&info);
-    if (!CreateProcessW(
-            /* lpApplicationName */ NULL,
-            /* lpCommandLine */ mutable_commandline.get(),
-            /* lpProcessAttributes */ NULL,
-            /* lpThreadAttributes */ NULL,
-            /* bInheritHandles */ TRUE,
-            /* dwCreationFlags */ CREATE_NO_WINDOW  // Don't create console
-                                                    // window
-                |
-                CREATE_NEW_PROCESS_GROUP  // So that Ctrl-Break isn't propagated
-                | CREATE_SUSPENDED  // So that it doesn't start a new job itself
-                | EXTENDED_STARTUPINFO_PRESENT | CREATE_UNICODE_ENVIRONMENT,
-            /* lpEnvironment */ env_map.ptr(),
-            /* lpCurrentDirectory */ cwd.empty() ? NULL : cwd.c_str(),
-            /* lpStartupInfo */ &info.StartupInfo,
-            /* lpProcessInformation */ &process_info)) {
-      DWORD err = GetLastError();
-      error_ = bazel::windows::MakeErrorMessage(
-          WSTR(__FILE__), __LINE__, L"CreateProcessW", cmd_sample, err);
-      return false;
-    }
-
-    pid_ = process_info.dwProcessId;
-    process_ = process_info.hProcess;
-    thread = process_info.hThread;
-
-    if (!AssignProcessToJobObject(job_, process_)) {
-      BOOL is_in_job = false;
-      if (IsProcessInJob(process_, NULL, &is_in_job) && is_in_job &&
-          !NestedJobsSupported()) {
-        // We are on a pre-Windows 8 system and the Bazel is already in a job.
-        // We can't create nested jobs, so just revert to TerminateProcess() and
-        // hope for the best. In batch mode, the launcher puts Bazel in a job so
-        // that will take care of cleanup once the command finishes.
-        CloseHandle(job_);
-        job_ = INVALID_HANDLE_VALUE;
-        CloseHandle(ioport_);
-        ioport_ = INVALID_HANDLE_VALUE;
-      } else {
-        DWORD err_code = GetLastError();
-        error_ = bazel::windows::MakeErrorMessage(
-            WSTR(__FILE__), __LINE__, L"nativeCreateProcess", wpath, err_code);
-        return false;
-      }
-    }
-
-    // Now that we put the process in a new job object, we can start executing
-    // it
-    if (ResumeThread(thread) == -1) {
-      DWORD err_code = GetLastError();
-      error_ = bazel::windows::MakeErrorMessage(
-          WSTR(__FILE__), __LINE__, L"nativeCreateProcess", wpath, err_code);
-      return false;
-    }
-
-    error_ = L"";
-    return true;
+    return proc_.Create(
+        wpath,
+        bazel::windows::GetJavaWstring(env, java_argv_rest),
+        bazel::windows::GetJavaWpath(env, java_cwd),
+        static_cast<void*>(env_map.ptr()),
+        stdin_process, stdout_process, stderr_process, &error_);
   }
 
   // Wait for this process to exit (or timeout).
   jint WaitFor(jlong timeout_msec) {
-    DWORD win32_timeout = timeout_msec < 0 ? INFINITE : timeout_msec;
-    jint result;
-    switch (WaitForSingleObject(process_, win32_timeout)) {
-      case WAIT_OBJECT_0:
-        result = kWaitSuccess;
-        break;
-
-      case WAIT_TIMEOUT:
-        result = kWaitTimeout;
-        break;
-
-      // Any other case is an error and should be reported back to Bazel.
-      default:
-        DWORD err_code = GetLastError();
-        error_ = bazel::windows::MakeErrorMessage(WSTR(__FILE__), __LINE__,
-                                                  L"NativeProcess:WaitFor",
-                                                  ToString(pid_), err_code);
-        return kWaitError;
-    }
-
-    if (stdin_ != INVALID_HANDLE_VALUE) {
-      CloseHandle(stdin_);
-      stdin_ = INVALID_HANDLE_VALUE;
-    }
-
-    // Ensure that the process is really terminated (if WaitForSingleObject
-    // above timed out, we have to explicitly kill it) and that it doesn't
-    // leave behind any subprocesses.
-    if (!Terminate()) {
-      return 2;
-    }
-
-    if (job_ != INVALID_HANDLE_VALUE) {
-      // Wait for the job object to complete, signalling that all subprocesses
-      // have exited.
-      DWORD CompletionCode;
-      ULONG_PTR CompletionKey;
-      LPOVERLAPPED Overlapped;
-      while (GetQueuedCompletionStatus(ioport_, &CompletionCode, &CompletionKey,
-                                       &Overlapped, INFINITE) &&
-             !((HANDLE)CompletionKey == job_ &&
-               CompletionCode == JOB_OBJECT_MSG_ACTIVE_PROCESS_ZERO)) {
-        // Still waiting...
-      }
-
-      CloseHandle(job_);
-      job_ = INVALID_HANDLE_VALUE;
-
-      CloseHandle(ioport_);
-      ioport_ = INVALID_HANDLE_VALUE;
-    }
-
-    // Fetch and store the exit code in case Bazel asks us for it later,
-    // because we cannot do this anymore after we closed the handle.
-    GetExitCode();
-
-    if (process_ != INVALID_HANDLE_VALUE) {
-      CloseHandle(process_);
-      process_ = INVALID_HANDLE_VALUE;
-    }
-
-    return result;
+    return static_cast<jint>(proc_.Wait(static_cast<long>(timeout_msec),
+                                                          &error_));
   }
 
   // Returns the exit code of the process if it has already exited. If the
   // process is still running, returns STILL_ACTIVE (= 259).
   jint GetExitCode() {
-    if (exit_code_ == STILL_ACTIVE) {
-      if (!GetExitCodeProcess(process_, &exit_code_)) {
-        DWORD err_code = GetLastError();
-        error_ = bazel::windows::MakeErrorMessage(WSTR(__FILE__), __LINE__,
-                                                  L"NativeProcess::GetExitCode",
-                                                  ToString(pid_), err_code);
-        return -1;
-      }
-    }
-
-    return exit_code_;
+    return static_cast<jint>(proc_.GetExitCode(&error_));
   }
 
-  jint GetPid() { return pid_; }
+  jint GetPid() { return static_cast<jint>(proc_.GetPid()); }
 
   jint WriteStdin(JNIEnv* env, jbyteArray java_bytes, jint offset,
                   jint length) {
     JavaByteArray bytes(env, java_bytes);
     if (offset < 0 || length <= 0 || offset > bytes.size() - length) {
       error_ = bazel::windows::MakeErrorMessage(
-          WSTR(__FILE__), __LINE__, L"NativeProcess:WriteStdin", ToString(pid_),
+          WSTR(__FILE__), __LINE__, L"NativeProcess:WriteStdin",
+          ToString(proc_.GetPid()),
           L"Array index out of bounds");
       return -1;
     }
@@ -595,7 +358,7 @@ class NativeProcess {
       DWORD err_code = GetLastError();
       error_ = bazel::windows::MakeErrorMessage(WSTR(__FILE__), __LINE__,
                                                 L"NativeProcess:WriteStdin",
-                                                ToString(pid_), err_code);
+                                                ToString(proc_.GetPid()), err_code);
       return -1;
     }
 
@@ -609,48 +372,7 @@ class NativeProcess {
 
   // Terminates this process (and subprocesses, if job objects are available).
   jboolean Terminate() {
-    static const UINT exit_code = 130;  // 128 + SIGINT, like on Linux
-
-    if (job_ != INVALID_HANDLE_VALUE) {
-      if (!TerminateJobObject(job_, exit_code)) {
-        DWORD err_code = GetLastError();
-        error_ = bazel::windows::MakeErrorMessage(WSTR(__FILE__), __LINE__,
-                                                  L"NativeProcess::Terminate",
-                                                  ToString(pid_), err_code);
-        return JNI_FALSE;
-      }
-    } else if (process_ != INVALID_HANDLE_VALUE) {
-      if (!TerminateProcess(process_, exit_code)) {
-        DWORD err_code = GetLastError();
-        std::wstring our_error = bazel::windows::MakeErrorMessage(
-            WSTR(__FILE__), __LINE__, L"NativeProcess::Terminate",
-            ToString(pid_), err_code);
-
-        // If the process exited, despite TerminateProcess having failed, we're
-        // still happy and just ignore the error. It might have been a race
-        // where the process exited by itself just before we tried to kill it.
-        // However, if the process is *still* running at this point (evidenced
-        // by its exit code still being STILL_ACTIVE) then something went
-        // really unexpectedly wrong and we should report that error.
-        if (GetExitCode() == STILL_ACTIVE) {
-          // Restore the error message from TerminateProcess - it will be much
-          // more helpful for debugging in case something goes wrong here.
-          error_ = our_error;
-          return JNI_FALSE;
-        }
-      }
-
-      if (WaitForSingleObject(process_, INFINITE) != WAIT_OBJECT_0) {
-        DWORD err_code = GetLastError();
-        error_ = bazel::windows::MakeErrorMessage(WSTR(__FILE__), __LINE__,
-                                                  L"NativeProcess::Terminate",
-                                                  ToString(pid_), err_code);
-        return JNI_FALSE;
-      }
-    }
-
-    error_ = L"";
-    return JNI_TRUE;
+    return proc_.Terminate(&error_) ? JNI_TRUE : JNI_FALSE;
   }
 
   // Return the last error as a human-readable string and clear it.
@@ -662,27 +384,11 @@ class NativeProcess {
   }
 
  private:
-  bool NestedJobsSupported() {
-    OSVERSIONINFOEX version_info;
-    version_info.dwOSVersionInfoSize = sizeof(version_info);
-    if (!GetVersionEx(reinterpret_cast<OSVERSIONINFO*>(&version_info))) {
-      return false;
-    }
-
-    return version_info.dwMajorVersion > 6 ||
-           (version_info.dwMajorVersion == 6 &&
-            version_info.dwMinorVersion >= 2);
-  }
-
   HANDLE stdin_;
   NativeOutputStream stdout_;
   NativeOutputStream stderr_;
-  HANDLE process_;
-  HANDLE job_;
-  HANDLE ioport_;
-  DWORD pid_;
-  DWORD exit_code_;
   std::wstring error_;
+  bazel::windows::WaitableProcess proc_;
 };
 
 // Ensure we can safely cast jlong to void*.
